move registration package back into meshctl
