+++ pkg/meshctl/registration/mocks/mock_registrant.go
