+++ pkg/meshctl/registration/registrant.go
